Move viewport settings to dedicated export
