+++ .env.example
@@ -5,3 +5,6 @@ GMAIL_APP_PASSWORD=xnglwckdxipdktto
 
 # Email account used by the contact form API
 CONTACT_EMAIL=maruoka@sally-inc.jp
+
+# Base URL of the website used for SEO metadata
+NEXT_PUBLIC_SITE_URL=http://localhost:3000
