+++ src/app/en/works/jilvain/page.tsx
@@ -4,6 +4,20 @@ import { ArrowLeft, ExternalLink } from "lucide-react"
 import { Button } from "@/components/ui/button"
 import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs"
 import { Card, CardContent } from "@/components/ui/card"
+import type { Metadata } from "next"
+
+export const metadata: Metadata = {
+  title: "JILVAIN - MARU",
+  description: "Details for the fantasy mystery 'JILVAIN' available on the UZU app.",
+  keywords: [
+    "JILVAIN",
+    "murder mystery",
+    "madamis",
+    "murder mystery app",
+    "UZU",
+    "MARU",
+  ],
+}
 
 export default function JilvainPage() {
   return (
