+++ src/app/en/works/shadow-code/page.tsx
@@ -4,6 +4,20 @@ import { ArrowLeft, ExternalLink } from "lucide-react"
 import { Button } from "@/components/ui/button"
 import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs"
 import { Card, CardContent } from "@/components/ui/card"
+import type { Metadata } from "next"
+
+export const metadata: Metadata = {
+  title: "SHADOW CODE - MARU",
+  description: "Details for the murder mystery 'SHADOW CODE' available on the UZU app.",
+  keywords: [
+    "SHADOW CODE",
+    "murder mystery",
+    "madamis",
+    "murder mystery app",
+    "UZU",
+    "MARU",
+  ],
+}
 
 export default function ShadowCodePage() {
   return (
